+++ src/helpers/utf8.js
@@ -45,6 +45,11 @@ module.exports.utf8 = function (/* start, end, handler */) {
   var args = this._helper_setArguments(['"', '"'], arguments, 'utf8')
   var handler = args[2]
 
+  // Special case: if end is not set, use the start value
+  var last = arguments[arguments.length-1]
+  if (arguments.length < 3 && (!last || typeof last === 'function'))
+    args[1] = args[0]
+
   function utf8Handler (data) {
     if (data.length < 2) {
       handler(data)